fix : lock timeout 문제 해결, facade 패턴으로 트랜잭션 분리
