+++ src/main/java/com/coffee_shop/coffeeshop/service/coupon/CouponService.java
@@ -1,7 +1,6 @@
 package com.coffee_shop.coffeeshop.service.coupon;
 
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.coffee_shop.coffeeshop.common.exception.BusinessException;
@@ -18,6 +17,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @Service
 public class CouponService {
+	private static final int SYNC_COUNT = 10;
 	private final CouponRepository couponRepository;
 
 	@Transactional
@@ -26,9 +26,14 @@ public class CouponService {
 		return savedCoupon.getId();
 	}
 
-	@Transactional(propagation = Propagation.REQUIRES_NEW)
+	@Transactional
 	public void syncCouponIssuedCount(Long couponId, Long increment) {
 		Coupon coupon = findCoupon(couponId);
+
+		if (!isDbSyncRequired(increment, coupon)) {
+			return;
+		}
+		
 		coupon.updateIssuedCount(increment);
 	}
 
@@ -36,5 +41,14 @@ public class CouponService {
 		return couponRepository.findById(couponId)
 			.orElseThrow(
 				() -> new BusinessException(ErrorCode.ENTITY_NOT_FOUND, "Coupon Not Found, 쿠폰 ID : " + couponId));
+	}
+
+	private static boolean isDbSyncRequired(Long increment, Coupon coupon) {
+		if (increment == null || increment <= 0L) {
+			log.warn("Coupon could not be updated because issuedCount is null or zero");
+			return false;
+		}
+
+		return increment % SYNC_COUNT == 0 || increment == coupon.getMaxIssueCount();
 	}
 }
